fixes to work in final prod env

+++ database/migrations/2021_02_26_120206_create_modules_table.php
@@ -17,7 +17,7 @@ class CreateModulesTable extends Migration
             $table->timestamps();
             $table->string('id')->primary();
             $table->string('title');
-            $table->integer('year');
+            $table->unsignedBigInteger('year');
         });
     }
